Refactored DatabaseFixture for Mss Created MssDbHelper
