+++ src/SqlServer.Tests/DatabaseFixture.cs
@@ -3,10 +3,8 @@
 public class DatabaseFixture : IAsyncLifetime
 {
     private MsSqlContainer? _mssContainer;
-    private readonly IParseTree _parseTree = new ParseTree(new NodeFactory(), new SqlTypes());
-    private readonly IPgParser _parser = new PgParser();
-    private readonly ITokenizerFactory _tokenizerFactory = new TokenizerFactory(new TokenFactory());
     private IDbContext? _mssDbContext;
+    private MssDbHelper? _mssDbHelper;
     private IConfiguration? _testConfiguration;
     private string? _connectionString;
 
@@ -18,6 +16,12 @@ public class DatabaseFixture : IAsyncLifetime
     {
         get => _mssDbContext ?? throw new ArgumentNullException(nameof(MssDbContext));
         set => _mssDbContext = value;
+    }
+
+    public MssDbHelper DbHelper
+    {
+        get => _mssDbHelper ?? throw new ArgumentNullException(nameof(MssDbHelper));
+        set => _mssDbHelper = value;
     }
 
     public IConfiguration TestConfiguration
@@ -45,147 +49,17 @@ public class DatabaseFixture : IAsyncLifetime
         _connectionString = TestConfiguration.GetConnectionString(Constants.Config.MssConnectionString);
 
         MssDbContext = new MssContext(TestConfiguration);
+        _mssDbHelper = new MssDbHelper(MssDbContext);
         await OneTimeSetupAsync();
     }
 
     private async Task OneTimeSetupAsync()
     {
-        var schemaExists = await ExecuteScalarAsync<int>($"select count(*) from sys.schemas where name = '{TestSchemaName}'");
+        var schemaExists = await DbHelper.ExecuteScalarAsync<int>($"select count(*) from sys.schemas where name = '{TestSchemaName}'");
         if (schemaExists == 0)
         {
-            await ExecuteNonQuery($"CREATE SCHEMA {TestSchemaName}");
+            await DbHelper.ExecuteNonQuery($"CREATE SCHEMA {TestSchemaName}");
         }
-    }
-
-    public async Task CreateTable(TableDefinition tableDefinition)
-    {
-        var sb = new StringBuilder();
-        sb.AppendLine($"create table [{tableDefinition.Header.Schema}].[{tableDefinition.Header.Name}] (");
-        var first = true;
-        foreach (var column in tableDefinition.Columns)
-        {
-            if (first)
-            {
-                first = false;
-            }
-            else
-            {
-                sb.AppendLine(",");
-            }
-
-            sb.Append($"    {column.Name} ");
-            sb.Append(column.GetTypeClause());
-            sb.Append(column.GetIdentityClause());
-            if (column.HasDefault)
-            {
-                var tokenizer = _tokenizerFactory.GetTokenizer();
-                tokenizer.Initialize(column.DefaultConstraint!.Definition);
-                tokenizer.GetNext();
-                var root = _parseTree.CreateExpression(tokenizer);
-                sb.Append(" default ");
-                sb.Append(_parser.Parse(tokenizer, root));
-            }
-            sb.Append(column.GetNullableClause());
-        }
-        sb.AppendLine(");");
-        await ExecuteNonQuery(sb.ToString());
-    }
-
-    public async Task DropTable(string tableName)
-    {
-        await DropTable("dbo", tableName);
-    }
-
-    public async Task DropTable(string schema, string tableName)
-    {
-        await ExecuteNonQuery($"DROP TABLE IF EXISTS [{schema}].[{tableName}];");
-    }
-
-    public async Task InsertIntoSingleColumnTable(
-        string tableName,
-        object? value,
-        SqlDbType? dbType = null)
-    {
-        await using var sqlConnection = new SqlConnection(MssConnectionString);
-        await sqlConnection.OpenAsync();
-        var sqlString = $"insert into [{tableName}] values (@Value);";
-        await using var sqlCommand = new SqlCommand(sqlString, sqlConnection);
-        if (dbType == null)
-        {
-            sqlCommand.Parameters.AddWithValue("@Value", value ?? DBNull.Value);
-        }
-        else
-        {
-            // Note: Since dbType is nullable, we have to cast to non-nullable
-            // otherwise we get the wrong SqlParameter constructor
-            // and dbType is treated as the value instead of the SqlDbType
-            var sqlParameter = new SqlParameter("@Value", (SqlDbType)dbType);
-            sqlParameter.Value = value ?? DBNull.Value;
-            sqlCommand.Parameters.Add(sqlParameter);
-        }
-        await sqlCommand.ExecuteNonQueryAsync();
-    }
-
-    public async Task CreateIndex(string tableName, IndexDefinition indexDefinition)
-    {
-        var sb = new StringBuilder();
-        sb.Append("create ");
-        if (indexDefinition.Header.IsUnique)
-        {
-            sb.Append("unique ");
-        }
-
-        if (indexDefinition.Header.Type == IndexType.Clustered)
-        {
-            sb.Append("clustered ");
-        }
-
-        sb.Append("index ");
-        sb.Append($"[{indexDefinition.Header.Name}] on ");
-        sb.Append($"[{tableName}] (");
-        var first = true;
-        foreach (var indexCol in indexDefinition.Columns)
-        {
-            if (first)
-            {
-                first = false;
-            }
-            else
-            {
-                sb.AppendLine(",");
-            }
-
-            sb.Append($"    {indexCol.Name} ");
-            if (indexCol.Direction == Direction.Descending)
-            {
-                sb.Append("desc");
-            }
-        }
-        sb.Append(")");
-
-        await ExecuteNonQuery(sb.ToString());
-    }
-
-    public async Task DropIndex(string tableName, string indexName)
-    {
-        var sqlString = $"if exists (select name from sys.indexes where object_id=object_id('{tableName}') and name = '{indexName}' drop index [{tableName}].[{indexName}];";
-        await ExecuteNonQuery(sqlString);
-    }
-
-    public async Task ExecuteNonQuery(string sqlString)
-    {
-        await using var sqlConnection = new SqlConnection(MssConnectionString);
-        await sqlConnection.OpenAsync();
-        await using var sqlCommand = new SqlCommand(sqlString, sqlConnection);
-        await sqlCommand.ExecuteNonQueryAsync();
-    }
-
-    public async Task<T?> ExecuteScalarAsync<T>(string sqlString)
-    {
-        await using var sqlConnection = new SqlConnection(MssConnectionString);
-        await sqlConnection.OpenAsync();
-        await using var sqlCommand = new SqlCommand(sqlString, sqlConnection);
-        return (T?)await sqlCommand.ExecuteScalarAsync();
     }
 
     public async Task DisposeAsync()
