+++ src/SqlServer.Tests/MssTestBase.cs
@@ -47,4 +47,47 @@ public abstract class MssTestBase
         var methodName = sf.GetMethod()?.Name ?? throw new InvalidOperationException("Method is null");
         return methodName.Length > 34 ? methodName[4..34] : methodName;
     }
+
+    public async Task CreateTable(TableDefinition tableDefinition)
+    {
+        await DbFixture.DbHelper.CreateTable(tableDefinition);
+    }
+
+    public async Task DropTable(string tableName)
+    {
+        await DropTable("dbo", tableName);
+    }
+
+    public async Task DropTable(string schema, string tableName)
+    {
+        await DbFixture.DbHelper.DropTable((schema, tableName));
+    }
+
+    public async Task InsertIntoSingleColumnTable(
+        string tableName,
+        object? value,
+        SqlDbType? dbType = null)
+    {
+        await DbFixture.DbHelper.InsertIntoSingleColumnTable(tableName, value, dbType);
+    }
+
+    public async Task CreateIndex(string tableName, IndexDefinition indexDefinition)
+    {
+        await DbFixture.DbHelper.CreateIndex(tableName, indexDefinition);
+    }
+
+    public async Task DropIndex(string tableName, string indexName)
+    {
+        await DbFixture.DbHelper.DropIndex(tableName, indexName);
+    }
+
+    public async Task ExecuteNonQuery(string sqlString)
+    {
+        await DbFixture.DbHelper.ExecuteNonQuery(sqlString);
+    }
+
+    public async Task<T?> ExecuteScalarAsync<T>(string sqlString)
+    {
+        return await DbFixture.DbHelper.ExecuteScalarAsync<T>(sqlString);
+    }
 }
