+++ src/SqlServer.Tests/SystemTables/MssTestGetTableHeader.cs
@@ -57,8 +57,8 @@ public class MssTestGetTableHeader(
 
     private async Task CreateTableAsync(string schema, string tableName)
     {
-        await DbFixture.DropTable(schema, tableName);
-        await DbFixture.ExecuteNonQuery(
+        await DropTable(schema, tableName);
+        await ExecuteNonQuery(
             $"CREATE TABLE [{schema}].[{tableName}](\r\n\t[Id] [bigint] IDENTITY(1,1) NOT NULL,\r\n\t[ExactNumBigInt] [bigint] NOT NULL)");
     }
 }
